+++ simulator/report.go
@@ -2,10 +2,10 @@ package simulator
 
 import "context"
 
-func (sim *SimulationSession) PrepareReportPerEpoch(ctx context.Context, slot uint64) error {
+func (sim *SimulationSession) PrepareReportPerEpoch(ctx context.Context, slot uint32) error {
 	return nil
 }
 
-func (sim *SimulationSession) PrepareReportPerSlot(ctx context.Context, slot uint64) error {
+func (sim *SimulationSession) PrepareReportPerSlot(ctx context.Context, slot uint32) error {
 	return nil
 }
